Prevention from lockdown

+++ Projekt/BuildingManagement/src/GUI/Admin/Remove.java
@@ -114,6 +114,10 @@ public class Remove extends Window {
                 errorWindow("Provide username!");
                 return;
             }
+            if (username.equals("admin")) {
+                errorWindow("You can't remove account 'admin'\nit is permanent account!");
+                return;
+            }
             errorCode = userDAO.deleteUserDB(username);
             if (errorCode == 0) messageWindow("User: " + username + " removed!");
             else errorWindow("Error: " + errorCode);
